+++ main/src/main/java/com/coggiri/main/domain/tag/repository/TagRepository.java
@@ -6,10 +6,10 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface TagRepository {
-    int addGroupTag(@Param("groupId")int groupId,@Param("tags") Tag tag);
-    int addTaskTag(@Param("taskId")int taskId,@Param("tags") Tag tag);
-    int deleteGroupTag(@Param("groupId") int groupId,@Param("tags") Tag tag);
-    int deleteTaskTag(@Param("taskId") int taskId,@Param("tags") Tag tag);
-    String[] getGroupTag(int groupId);
-    String[] getTaskTag(int taskId);
+    int addGroupTag(@Param("groupId")Long groupId,@Param("tags") Tag tag);
+    int addTaskTag(@Param("taskId")Long taskId,@Param("tags") Tag tag);
+    int deleteGroupTag(@Param("groupId") Long groupId,@Param("tags") Tag tag);
+    int deleteTaskTag(@Param("taskId") Long taskId,@Param("tags") Tag tag);
+    String[] getGroupTag(Long groupId);
+    String[] getTaskTag(Long taskId);
 }
